Cleanup before getting in
