+++ concept/type_/mod.rs
@@ -31,6 +31,7 @@ pub mod relation_type;
 pub mod role_type;
 pub mod type_cache;
 pub mod type_manager;
+mod storage_source;
 
 pub trait TypeAPI<'a>: ConceptAPI<'a> + Sized + Clone {
     fn vertex<'this>(&'this self) -> TypeVertex<'this>;
